+++ src/main/java/com/crowdin/cli/commands/actions/PreTranslateAction.java
@@ -74,11 +74,16 @@ class PreTranslateAction implements NewAction<PropertiesWithFiles, ProjectClient
 
         List<String> languages = this.prepareLanguageIds(project);
         List<Long> fileIds = this.prepareFileIds(out, properties, project);
+
+        if (fileIds == null || fileIds.isEmpty()) {
+            throw new RuntimeException(String.format(RESOURCE_BUNDLE.getString("error.no_files_found_for_pre_translate")));
+        }
+
         ApplyPreTranslationRequest request = RequestBuilder.applyPreTranslation(
             languages, fileIds, method, engineId, autoApproveOption,
             duplicateTranslations, translateUntranslatedOnly, translateWithPerfectMatchOnly);
-        this.applyPreTranslation(out, client, request);
 
+        this.applyPreTranslation(out, client, request);
     }
 
     private List<String> prepareLanguageIds(CrowdinProjectInfo projectInfo) {
@@ -139,19 +144,33 @@ class PreTranslateAction implements NewAction<PropertiesWithFiles, ProjectClient
             .collect(Collectors.toList());
     }
 
-    private PreTranslationStatus applyPreTranslation(Outputter out, ProjectClient client,ApplyPreTranslationRequest request) {
-        return ConsoleSpinner.execute(out, "message.spinner.pre_translate", "error.spinner.pre_translate", this.noProgress, this.plainView, () -> {
-            PreTranslationStatus preTranslationStatus = client.startPreTranslation(request);
+    private void applyPreTranslation(Outputter out, ProjectClient client, ApplyPreTranslationRequest request) {
+        ConsoleSpinner.execute(
+            out,
+            "message.spinner.pre_translate",
+            "error.spinner.pre_translate",
+            this.noProgress,
+            this.plainView,
+            () -> {
+                PreTranslationStatus preTranslationStatus = client.startPreTranslation(request);
 
-            while (!preTranslationStatus.getStatus().equalsIgnoreCase("finished")) {
-                ConsoleSpinner.update(
-                    String.format(RESOURCE_BUNDLE.getString("message.spinner.pre_translate_percents"),
-                        Math.toIntExact(preTranslationStatus.getProgress())));
-                Thread.sleep(100);
-                preTranslationStatus = client.checkPreTranslation(preTranslationStatus.getIdentifier());
+                while (!preTranslationStatus.getStatus().equalsIgnoreCase("finished")) {
+                    ConsoleSpinner.update(
+                            String.format(RESOURCE_BUNDLE.getString("message.spinner.pre_translate_percents"),
+                                    Math.toIntExact(preTranslationStatus.getProgress())));
+                    Thread.sleep(100);
+
+                    preTranslationStatus = client.checkPreTranslation(preTranslationStatus.getIdentifier());
+
+                    if (preTranslationStatus.getStatus().equalsIgnoreCase("failed")) {
+                        throw new RuntimeException();
+                    }
+                }
+
+                ConsoleSpinner.update(String.format(RESOURCE_BUNDLE.getString("message.spinner.pre_translate_done"), 100));
+
+                return preTranslationStatus;
             }
-            ConsoleSpinner.update(String.format(RESOURCE_BUNDLE.getString("message.spinner.pre_translate_done"), 100));
-            return preTranslationStatus;
-        });
+        );
     }
 }