+++ src/main/java/com/crowdin/cli/commands/actions/UploadSourcesAction.java
@@ -307,6 +307,7 @@ class UploadSourcesAction implements NewAction<PropertiesWithFiles, ProjectClien
             SpreadsheetFileImportOptions importOptions = new SpreadsheetFileImportOptions();
             importOptions.setFirstLineContainsHeader(fileBean.getFirstLineContainsHeader());
             importOptions.setScheme(PropertiesBeanUtils.getSchemeObject(fileBean.getScheme()));
+            importOptions.setImportTranslations(fileBean.getImportTranslations());
             return importOptions;
         } else if (isXml(sourceFile)) {
             XmlFileImportOptions importOptions = new XmlFileImportOptions();