+++ src/main/java/com/crowdin/cli/commands/functionality/RequestBuilder.java
@@ -134,13 +134,14 @@ public class RequestBuilder {
     }
 
     public static ExportProjectTranslationRequest exportProjectTranslation(
-        String format, Boolean skipUntranslatedStrings, Boolean skipUntranslatedFiles, Integer exportWithMinApprovalsCount
+        String format, Boolean skipUntranslatedStrings, Boolean skipUntranslatedFiles, Integer exportWithMinApprovalsCount, Boolean exportStringsThatPassedWorkflow
     ) {
         ExportProjectTranslationRequest request = new ExportProjectTranslationRequest();
         request.setFormat(format);
         request.setSkipUntranslatedStrings(skipUntranslatedStrings);
         request.setSkipUntranslatedFiles(skipUntranslatedFiles);
         request.setExportWithMinApprovalsCount(exportWithMinApprovalsCount);
+        request.setExportStringsThatPassedWorkflow(exportStringsThatPassedWorkflow);
         return request;
     }
 
@@ -156,7 +157,19 @@ public class RequestBuilder {
         copy.setSkipUntranslatedFiles(request.getSkipUntranslatedFiles());
         copy.setExportApprovedOnly(request.getExportApprovedOnly());
         copy.setExportWithMinApprovalsCount(request.getExportWithMinApprovalsCount());
+        copy.setExportStringsThatPassedWorkflow(request.getExportStringsThatPassedWorkflow());
         return copy;
+    }
+
+    public static CrowdinTranslationCraeteProjectPseudoBuildForm crowdinTranslationCreateProjectPseudoBuildForm(
+        long branchId, Boolean pseudo, Integer lengthCorrection, String prefix, String suffix, CharTransformation charTransformation
+    ) {
+        CrowdinTranslationCraeteProjectPseudoBuildForm request
+            = crowdinTranslationCreateProjectPseudoBuildForm(pseudo, lengthCorrection, prefix, suffix, charTransformation);
+
+        request.setBranchId(branchId);
+
+        return request;
     }
 
     public static CrowdinTranslationCraeteProjectPseudoBuildForm crowdinTranslationCreateProjectPseudoBuildForm(
@@ -181,6 +194,7 @@ public class RequestBuilder {
         requestCopy.setSkipUntranslatedFiles(request.getSkipUntranslatedFiles());
         requestCopy.setExportApprovedOnly(request.getExportApprovedOnly());
         requestCopy.setExportWithMinApprovalsCount(request.getExportWithMinApprovalsCount());
+        requestCopy.setExportStringsThatPassedWorkflow(request.getExportStringsThatPassedWorkflow());
         return requestCopy;
     }