+++ src/main/java/com/crowdin/cli/properties/PropertiesBuilder.java
@@ -112,6 +112,8 @@ public abstract class PropertiesBuilder<T extends Properties, P extends Params> 
 
     public static final String EXPORT_APPROVED_ONLY = "export_only_approved";
 
+    public static final String EXPORT_STRINGS_THAT_PASSED_WORKFLOW = "export_strings_that_passed_workflow";
+
     public static final String EXCLUDED_TARGET_LANGUAGES = "excluded_target_languages";
 
     public static final String CUSTOM_SEGMENTATION = "custom_segmentation";
@@ -121,6 +123,8 @@ public abstract class PropertiesBuilder<T extends Properties, P extends Params> 
     public static final String IGNORE_HIDDEN_FILES = "ignore_hidden_files";
 
     public static final String CONFIG_FILE_PATH = "config_file_path";
+
+    public static final String IMPORT_TRANSLATIONS = "import_translations";
 
     private Outputter out;
     private Map<String, Object> configFileParams;