+++ CHANGELOG.md
@@ -1,6 +1,56 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [3.9.0]
+
+### Added
+
+- Add downloading multiple specified languages ([#493](https://github.com/crowdin/crowdin-cli/pull/493))
+- Add '--dryrun' mode to the 'download sources' command ([#485](https://github.com/crowdin/crowdin-cli/pull/485))
+- Add the `export_string_that_passed_workflow` export option support ([#487](https://github.com/crowdin/crowdin-cli/pull/487))
+- Add Unit tests for `StatusAction.java` ([#494](https://github.com/crowdin/crowdin-cli/pull/494))
+
+### Updated
+
+- Improve async progress status checking for some commands ([#478](https://github.com/crowdin/crowdin-cli/pull/478))
+- Make lint throw error with no config file ([#483](https://github.com/crowdin/crowdin-cli/pull/483))
+
+### Fixed
+
+- Fixed regex to work correctly for Windows machines ([#484](https://github.com/crowdin/crowdin-cli/pull/484))
+
+## [3.8.1]
+
+### Fixed
+
+- Fix infinite build progress check in case of the build was failed ([#476](https://github.com/crowdin/crowdin-cli/pull/476))
+
+## [3.8.0]
+
+### Added
+
+- Download reviewed sources command ([#471](https://github.com/crowdin/crowdin-cli/pull/471))
+- `--fail-if-incomplete` option for status commands ([#466](https://github.com/crowdin/crowdin-cli/pull/466))
+- `branch` parameter support for Pseudo-Localization download ([#472](https://github.com/crowdin/crowdin-cli/pull/472))
+
+## [3.7.10]
+
+### Added
+
+- `import_translations` option support for spreadsheet files ([#463](https://github.com/crowdin/crowdin-cli/pull/463))
+
+### Updated
+
+- Improved error handling for the `pre-translate` command ([#461](https://github.com/crowdin/crowdin-cli/pull/461))
+
+## [3.7.9]
+
+### Updated
+
+- Support language ID in list languages command ([#454](https://github.com/crowdin/crowdin-cli/pull/454))
+- Update jDeploy ([#437](https://github.com/crowdin/crowdin-cli/pull/437))
+- Dependencies update ([#442](https://github.com/crowdin/crowdin-cli/pull/442))
+
 ## [3.7.8]
 
 ### Updated